requirements.txt
Adafruit-Blinka==8.20.1
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-neopixel==6.3.9
adafruit-circuitpython-pixelbuf==2.0.2
adafruit-circuitpython-requests==2.0.1
adafruit-circuitpython-typing==1.9.4
Adafruit-PlatformDetect==3.49.0
Adafruit-PureIO==1.1.11
awscrt==0.21.1
awsiot==0.1.3
awsiotsdk==1.22.0
boto3==1.34.155
botocore==1.34.155
click==8.1.7
jmespath==1.0.1
pyftdi==0.55.0
pyserial==3.5
python-dateutil==2.8.2
pyusb==1.2.1
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
fake_rpi==0.7.1 # use this for non-Raspi testing, instead of RPi.GPIO
six==1.16.0
sysv-ipc==1.1.0
typing_extensions==4.7.1
urllib3==1.26.16
keyboard